Refactor imports and clean up code across multiple modules
- Removed unused imports and reorganized import statements for clarity.
- Updated type hints to use `Optional` with the new syntax.
- Cleaned up whitespace and formatting inconsistencies.
- Enhanced readability by grouping related imports together.
- Improved the handling of asyncio tasks in the Discord voice channel repository.
- Adjusted assertions in tests for consistency and clarity.
- Ensured all modules adhere to the same import style and structure.

+++ src/infrastructure/runtime_observability.py
@@ -116,5 +116,5 @@ class InMemoryBotRuntimeTelemetry:
         if not values:
             return None
         sorted_values = sorted(values)
-        index = int(round((len(sorted_values) - 1) * percentile))
+        index = round((len(sorted_values) - 1) * percentile)
         return sorted_values[index]
